start-cluster.sh: allow overriding kubernetes dir instead of run.sh
Now start-cluster.sh takes an optional argument, which is a path
to a directory containing kubernetes config directory for bringing
up a conservator instance. If the path is not supplied, then the
kubernetes config directory will be pulled from a docker image
(like it was before).

The mechanism to override just kubernetes/kind/run.sh is being
removed, since now you can pass in a modified kubernetes/ dir instead.

+++ test/integration/cluster/start-cluster.sh
@@ -26,9 +26,16 @@ env aws ecr get-login-password --region us-east-1 \
                | docker login --username AWS --password-stdin $AWS_DOMAIN
 docker pull $AWS_DOMAIN/conservator_webapp:prod -q
 echo "** Create Conservator Cluster"
-id=$(docker create $AWS_DOMAIN/conservator_webapp:prod)
-docker cp $id:/home/centos/flirmachinelearning/docker/kubernetes/ kubernetes/
-docker rm -v $id
-cp override-kind-run.sh kubernetes/kind/run.sh #@@@@@@ FIXME just until updated run.sh gets into prod
+rm -rf ./kubernetes
+if [ -d "$1" ] ; then
+  echo "-- Copy external kubernetes configs from '$1'"
+  cp -r $1 kubernetes
+else
+  echo "-- Copy kubernetes configs from docker image"
+  id=$(docker create $AWS_DOMAIN/conservator_webapp:prod)
+  docker cp $id:/home/centos/flirmachinelearning/docker/kubernetes/ kubernetes/
+  docker rm -v $id
+fi
+#cp override-kind-run.sh kubernetes/kind/run.sh #@@@@@@ FIXME just until updated run.sh gets into prod
 cd kubernetes
 ./kind/run.sh ../kind-init.env
